feat(builders): :sparkles: new workspace builder for grouped releases
