+++ libs/builders/project.json
@@ -59,5 +59,6 @@
       }
     }
   },
-  "tags": []
+  "tags": [],
+  "implicitDependencies": ["workspace-release-image"]
 }
