+++ libs/builders/src/semantic-release/base-schema.ts
@@ -1,0 +1,13 @@
+import { JsonObject } from '@angular-devkit/core';
+
+import { BranchSpecJson } from './models';
+
+export interface SemanticReleaseBaseSchema extends JsonObject {
+  branches: BranchSpecJson[];
+  changelog: boolean;
+  dryRun: boolean;
+  force: boolean;
+  github: boolean;
+  npm: boolean;
+  releaseCommitMessage: string;
+}
