+++ libs/builders/src/semantic-release/project-schema.ts
@@ -1,14 +1,5 @@
-import { JsonObject } from '@angular-devkit/core';
+import { SemanticReleaseBaseSchema } from './base-schema';
 
-import { BranchSpecJson } from './models';
-
-export interface SemanticReleaseProjectSchema extends JsonObject {
-  dryRun: boolean;
-  force: boolean;
+export interface SemanticReleaseProjectSchema extends SemanticReleaseBaseSchema {
   mode: 'independent' | 'sync';
-  branches: Array<BranchSpecJson | string>;
-  releaseCommitMessage: string;
-  changelog: boolean;
-  npm: boolean;
-  github: boolean;
 }
