+++ libs/builders/src/semantic-release/workspace-schema.ts
@@ -1,14 +1,5 @@
-import { JsonObject } from '@angular-devkit/core';
+import { SemanticReleaseBaseSchema } from './base-schema';
 
-import { BranchSpecJson } from './models';
-
-export interface SemanticReleaseWorkspaceSchema extends JsonObject {
-  dryRun: boolean;
-  force: boolean;
+export interface SemanticReleaseWorkspaceSchema extends SemanticReleaseBaseSchema {
   tag: string;
-  branches: Array<BranchSpecJson | string>;
-  releaseCommitMessage: string;
-  changelog: boolean;
-  npm: boolean;
-  github: boolean;
 }
