+++ libs/image-optimizer/project.json
@@ -27,14 +27,6 @@
         "jestConfig": "libs/image-optimizer/jest.config.ts",
         "codeCoverage": true
       }
-    },
-    "release": {
-      "executor": "./dist/libs/builders:semantic-release",
-      "configurations": {
-        "local": {
-          "force": true
-        }
-      }
     }
   },
   "tags": ["release:image"]
